Change the aux field type to long instead of int so that it is large enough to hold a man-sized pointer.

+++ gnu/lib/libdialog/dialog.h
@@ -67,7 +67,7 @@ typedef struct _dmenu_item {
   void (*selected)(struct _dmenu_item *self, int is_selected);
   void *data;
   char lbra, mark, rbra;
-  int aux;
+  long aux;
 } dialogMenuItem;
 
 #define VERSION "0.4"
